+++ tests/test_filters.py
@@ -4,7 +4,7 @@ Tests write output (should it return True or False or write)
 """
 from atropos.commands.trim.filters import (
     NContentFilter, DISCARD, KEEP, SingleWrapper, PairedWrapper)
-from atropos.io.seqio import Sequence
+from atropos.io import Sequence
 
 def test_ncontentfilter():
     # third parameter is True if read should be discarded