+++ tests/test_modifiers.py
@@ -2,7 +2,7 @@
 from atropos.adapters import *
 from atropos.align import MatchInfo
 from atropos.commands.trim.modifiers import *
-from atropos.io.seqio import Sequence
+from atropos.io import Sequence
 from atropos.util import reverse_complement as rc
 
 DUMMY_ADAPTER = Adapter("ACGT", FRONT)